Lock CPU hotplug while processing results
le it's unlikely for a CPU to be unplugged in our testing environment,
we should still follow proper synchronization practices. This change
adds CPU hotplug locking around the results processing code to ensure
correctness in all scenarios.

The lock prevents CPU hotplug operations from interfering with the
processing, maintaining data consistency even if a CPU were to be
removed during this critical section.It is unprobable a CPU be
unplugged in our testing environment, but it is better to do things
in the right way.

Signed-off-by: Wander Lairson Costa <wander@redhat.com>